Add login to admin page
